+++ app/users/dashboard/workforce/leaves/policies/page.tsx
@@ -14,10 +14,10 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
-import { Badge } from '@/components/ui/badge';
-import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
+} from '@/components/shadcn/card';
+import { Button } from '@/components/shadcn/button';
+import { Badge } from '@/components/shadcn/badge';
+import { Alert, AlertDescription, AlertTitle } from '@/components/shadcn/alert';
 import {
   Settings,
   ArrowLeft,
